requirements.txt
asgiref==3.8.1
click==8.1.8
decorator==5.2.1
diff-match-patch==20241021
Django==5.1.7
django-cors-headers==4.7.0
django-import-export==4.3.7
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
numpy==2.2.3
pillow==10.4.0
proglog==0.1.10
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.2.1
rq==2.1.0
sqlparse==0.5.3
tablib==3.8.0
tqdm==4.67.1
